+++ xrpl/transaction/nftoken_create_offer.go
@@ -21,8 +21,6 @@ const (
 // **********************************
 
 var (
-	// ErrOwnerAccountConflict is returned when the owner is the same as the account.
-	ErrOwnerAccountConflict = errors.New("owner must be different from the account")
 	// ErrOwnerPresentForSellOffer is returned when the owner is present for a sell offer.
 	ErrOwnerPresentForSellOffer = errors.New("owner must not be present for a sell offer")
 	// ErrOwnerNotPresentForBuyOffer is returned when the owner is not present for a buy offer.
@@ -80,18 +78,18 @@ func (n *NFTokenCreateOffer) Flatten() FlatTransaction {
 	flattened["TransactionType"] = "NFTokenCreateOffer"
 
 	if n.Owner != "" {
-		flattened["Owner"] = n.Owner
+		flattened["Owner"] = n.Owner.String()
 	}
 
-	flattened["NFTokenID"] = n.NFTokenID
-	flattened["Amount"] = n.Amount
+	flattened["NFTokenID"] = n.NFTokenID.String()
+	flattened["Amount"] = n.Amount.Flatten()
 
 	if n.Expiration != 0 {
 		flattened["Expiration"] = n.Expiration
 	}
 
 	if n.Destination != "" {
-		flattened["Destination"] = n.Destination
+		flattened["Destination"] = n.Destination.String()
 	}
 
 	return flattened