+++ CHANGELOG.md
@@ -16,8 +16,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### xrpl
 
+- Added `AMMClawback` transaction type.
 - Adds `MPTokenAuthorize`, `MPTokenIssuanceCreate`, `MPTokenIssuanceDestroy`, `MPTokenIssuanceSet` transactions. It also adds the `types.Holder`, `types.AssetScale`, `types.MPTokenMetadata` and `types.TransferFee` types to represent the holder of the token, the asset scale, the metadata and the transfer fee of the token respectively.
 - Adds NFTokenMintOffer support by adding `Amount`, `Expiration`, and `Destination` fields to `NFTokenMint` transaction. Also add `NFTokenMintMetadata` struct to handle transaction metadata with `nftoken_id` and `offer_id` fields.
+- Adds `NFTokenModify` transaction type.
+
+## Changed
+
+### binary-codec
+
+- Refactor on `Issue` codec type to support `Currency` and `Issuer` fields.
+
+### Dependencies
+
+- Bumped Go version to 1.23.0.
+
+## Fixed
+
+### xrpl
+
+- Fixed some flattening fields with the `Flatten` function for `NFTokenMint`, `NFTokenCancel`, `NFTokenCreate`, `NFTokenBurn`
 
 ## [v0.1.10]